move SDC generator for routing modules to an independent source file
